# apps/architectural-service/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "Architectural Service API",
...
    "contact": {
      "name": "DesignSynapse Support",
      "url": "https://designsynapse.com/support",
      "email": "support@designsynapse.com"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://designsynapse.com/license"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/designs/{design_id}/materials": {
      "post": {
        "tags": [
          "materials"
        ],
        "summary": "Add material specification",
        "description": "Add a material specification to a design",
        "operationId": "add_material_specification_api_v1_designs__design_id__materials_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "design_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Design Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MaterialSpecificationRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaterialSpecificationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "materials"
        ],
        "summary": "List material specifications",
        "description": "List all material specifications for a design",
        "operationId": "list_material_specifications_api_v1_designs__design_id__materials_get",
        "parameters": [
          {
            "name": "design_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Design Id"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Filter by material category",
              "title": "Category"
            },
            "description": "Filter by material category"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MaterialSpecificationResponse"
                  },
                  "title": "Response List Material Specifications Api V1 Designs  Design Id  Materials Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/materials/search": {
      "get": {
        "tags": [
          "materials"
        ],
        "summary": "Search materials",
        "description": "Search for materials across all designs with vendor information",
        "operationId": "search_materials_api_v1_materials_search_get",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Search query for materials",
              "title": "Query"
            },
            "description": "Search query for materials"
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Filter by material category",
              "title": "Category"
            },
            "description": "Filter by material category"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "description": "Maximum number of results",
              "default": 50,
              "title": "Limit"
            },
            "description": "Maximum number of results"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MaterialSpecificationResponse"
                  },
                  "title": "Response Search Materials Api V1 Materials Search Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MaterialCategory": {
        "type": "string",
        "enum": [
          "structural",
          "finishes",
          "mechanical",
          "electrical",
          "plumbing",
          "insulation",
          "roofing",
          "glazing",
          "doors_windows",
          "hardware"
        ],
        "title": "MaterialCategory",
        "description": "Material categories."
      },
      "MaterialProperties": {
        "properties": {
          "type": {
            "type": "string",
            "maxLength": 100,
            "title": "Type",
            "description": "Material type"
          },
          "grade": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Grade",
            "description": "Material grade"
          },
          "dimensions": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Dimensions",
            "description": "Material dimensions"
          },
          "finish": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Finish",
            "description": "Material finish"
          },
          "properties": {
            "additionalProperties": true,
            "type": "object",
            "title": "Properties",
            "description": "Additional properties"
          }
        },
        "type": "object",
        "required": [
          "type"
        ],
        "title": "MaterialProperties",
        "description": "Material properties."
      },
      "MaterialSpecificationRequest": {
        "properties": {
          "category": {
            "$ref": "#/components/schemas/MaterialCategory",
            "description": "Material category"
          },
          "properties": {
            "$ref": "#/components/schemas/MaterialProperties",
            "description": "Material properties"
          },
          "design_elements": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Design Elements",
            "description": "Design elements using this material"
          }
        },
        "type": "object",
        "required": [
          "category",
          "properties"
        ],
        "title": "MaterialSpecificationRequest",
        "description": "Request schema for material specification.",
        "example": {
          "category": "structural",
          "design_elements": [],
          "properties": {
            "type": "Steel",
            "grade": "A992",
            "dimensions": "W12x26",
            "finish": "Painted",
            "properties": {
              "yield_strength": "50 ksi",
              "tensile_strength": "65 ksi"
            }
          }
        }
      },
      "MaterialSpecificationResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "Material specification ID"
          },
          "design_id": {
            "type": "string",
            "format": "uuid",
            "title": "Design Id",
            "description": "Associated design ID"
          },
          "category": {
            "$ref": "#/components/schemas/MaterialCategory",
            "description": "Material category"
          },
          "properties": {
            "$ref": "#/components/schemas/MaterialProperties",
            "description": "Material properties"
          },
          "vendor_info": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VendorInfo"
              },
              {
                "type": "null"
              }
            ],
            "description": "Vendor information"
          },
          "cost_estimate": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost Estimate",
            "description": "Cost estimate"
          },
          "design_elements": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Design Elements",
            "description": "Associated design elements"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Creation timestamp"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "Last update timestamp"
          }
        },
        "type": "object",
        "required": [
          "id",
          "design_id",
          "category",
          "properties",
          "created_at",
          "updated_at"
        ],
        "title": "MaterialSpecificationResponse",
        "description": "Response schema for material specification."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VendorInfo": {
        "properties": {
          "vendor_id": {
            "type": "string",
            "format": "uuid",
            "title": "Vendor Id",
            "description": "Vendor ID"
          },
          "vendor_name": {
            "type": "string",
            "title": "Vendor Name",
            "description": "Vendor name"
          },
          "product_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Product Code",
            "description": "Vendor product code"
          },
          "availability": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Availability",
            "description": "Availability status"
          },
          "lead_time_days": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lead Time Days",
            "description": "Lead time in days"
          }
        },
        "type": "object",
        "required": [
          "vendor_id",
          "vendor_name"
        ],
        "title": "VendorInfo",
        "description": "Vendor information for material."
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}